Move Shower out of resources & into code; make loading facility generic
